requirements.txt
langchain
langchain-community 
langchain-google-genai
langchain-chroma
faiss-cpu
nltk==3.9.1
unstructured
python-dotenv
numpy
flask
-e .